+++ public/assets/scripts/700.kernel.js
@@ -1,5 +1,5 @@
 /*!
- * Winbows11 - c753931b0a93507d
+ * Winbows11 - 3a727ea143e2d02a
  * Copyright (c) Microhard 2026
  * Github : Siyu1017/winbows11
  */